Renamed components dir, used interface
