+++ template/app/src/client/landing-page/components/FAQ.tsx
@@ -1,4 +1,4 @@
-type FAQ = {
+interface FAQ {
 	id: number;
 	question: string;
 	answer: string;
