+++ template/app/src/client/landing-page/components/Features.tsx
@@ -1,4 +1,4 @@
-type Feature = {
+interface Feature {
 	name: string;
 	description: string;
 	icon: string;
