+++ template/app/src/client/landing-page/components/Footer.tsx
@@ -1,4 +1,4 @@
-type NavigationItem = {
+interface NavigationItem {
 	name: string;
 	href: string;
 };
