+++ template/app/src/client/landing-page/components/Hero.tsx
